refactor: extract FilterCheckboxGroup, PriceRangeFilter, and ActiveFilters
