+++ src/components/ProductFilters/FilterCheckboxGroup.tsx
@@ -1,0 +1,69 @@
+import {
+  AccordionItem,
+  AccordionButton,
+  AccordionPanel,
+  Box,
+  Heading,
+  AccordionIcon,
+  CheckboxGroup,
+  Stack,
+  Checkbox,
+  Text,
+  Flex,
+  Spinner,
+} from '@chakra-ui/react';
+import { getLocalizedText } from '../../utils/localization';
+import { FilterState } from '../../features/filter/filterSlice';
+
+interface Props {
+  label: string;
+  filterKey: keyof Pick<FilterState, 'brand' | 'color' | 'size' | 'materials'>;
+  options: string[];
+  values: string[];
+  loading: boolean;
+  onChange: (values: string[]) => void;
+}
+
+const FilterCheckboxGroup = ({
+  label,
+  options,
+  values,
+  loading,
+  onChange,
+}: Props) => (
+  <AccordionItem>
+    <h2>
+      <AccordionButton>
+        <Box flex="1" textAlign="left">
+          <Heading size="sm">{label}</Heading>
+        </Box>
+        <AccordionIcon />
+      </AccordionButton>
+    </h2>
+    <AccordionPanel pb={4}>
+      {loading ? (
+        <Flex justify="center" py={2}>
+          <Spinner size="sm" />
+        </Flex>
+      ) : (
+        <CheckboxGroup value={values} onChange={onChange}>
+          <Stack spacing={1}>
+            {options.length > 0 ? (
+              options.map((value) => (
+                <Checkbox key={value} value={value}>
+                  {getLocalizedText({ en: value })}
+                </Checkbox>
+              ))
+            ) : (
+              <Text fontSize="sm" color="gray.500">
+                No {label.toLowerCase()} available
+              </Text>
+            )}
+          </Stack>
+        </CheckboxGroup>
+      )}
+    </AccordionPanel>
+  </AccordionItem>
+);
+
+export default FilterCheckboxGroup;
