+++ src/pages/ProductsPage.tsx
@@ -26,7 +26,7 @@ import {
 } from '../features/category/categorySlice';
 import { AppDispatch, RootState } from '../store/store';
 import ProductCard from '../components/ProductCard';
-import ProductFilters from '../components/ProductFilters';
+import ProductFilters from '../components/ProductFilters/ProductFilters.tsx';
 import { getLocalizedText } from '../utils/localization';
 
 const ProductsPage = () => {
